Added campaign id as a header, and used moment to format the date
